Support both exact and inclusive matching
Previously `assert_json_eq` would do an inclusive match, which allowed
addtional data in `actual`, but not in `expected`.

This changes `assert_json_eq` to instead perform an *exact* match, and
adds `assert_json_include` which performs an inclusive match.

This library is still young so I decided that this breaking change was
worth it.